cleanup(storage)!: move `KeyAes256`
