+++ src/storage/src/model.rs
@@ -1,0 +1,24 @@
+// Copyright 2025 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! The messages and enums that are part of this client library.
+
+// Re-export all generated types
+pub use crate::control::model::*;
+
+// Custom types used in the hand-crafted code. We do not expect this name to
+// conflict with generated types, that would require a `RequestHelpers` message
+// with nested enums or messages. If we ever get a conflict, we would configure
+// sidekick to rename the generated types.
+pub mod request_helpers;
