+++ src/storage/src/storage/read_object.rs
@@ -16,6 +16,7 @@ use super::client::*;
 use super::*;
 use crate::error::{RangeError, ReadError};
 use crate::model::ObjectChecksums;
+use crate::model::request_helpers::KeyAes256;
 use crate::read_resume_policy::ReadResumePolicy;
 use crate::storage::checksum::{
     ChecksumEngine,
@@ -288,7 +289,7 @@ where
     ///
     /// Example:
     /// ```
-    /// # use google_cloud_storage::{builder::storage::KeyAes256, client::Storage};
+    /// # use google_cloud_storage::{model::request_helpers::KeyAes256, client::Storage};
     /// # async fn sample(client: &Storage) -> anyhow::Result<()> {
     /// let key: &[u8] = &[97; 32];
     /// let response = client
@@ -912,9 +913,10 @@ mod resume_tests;
 
 #[cfg(test)]
 mod tests {
-    use super::client::tests::{create_key_helper, test_builder, test_inner_client};
+    use super::client::tests::{test_builder, test_inner_client};
     use super::*;
     use crate::error::ChecksumMismatch;
+    use crate::model::request_helpers::{KeyAes256, tests::create_key_helper};
     use futures::TryStreamExt;
     use httptest::{Expectation, Server, matchers::*, responders::status_code};
     use std::collections::HashMap;
